pyproject.toml
[tool.poetry]
name = "dz_fastapi"
version = "0.1.0"
description = "Project Dragonzap backend"
authors = ["Alexander Ermolenko <alex.ermolenko@gmail.com>"]
readme = "README.md"
packages = [
    { include = "dz_fastapi" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-multipart = "^0.0.9"
pillow = "^10.4.0"
aiofiles = "^24.1.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
sqlalchemy-utils = "^0.41.2"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
imap-tools = "^1.7.4"
apscheduler = "^3.11.0"
flake8 = "^7.1.1"
flake8-isort = "^6.1.1"
black = "^24.10.0"
xlrd = "^2.0.1"
rarfile = "^4.2"
libarchive-c = "^5.2"
plotly = "^6.0.1"
requests = "^2.32.3"
aiohttp = "^3.12.11"
xlsxwriter = "^3.2.3"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
